add notification service and tracing

+++ clients/src/main/java/com/aweperi/clients/fraud/FraudClient.java
@@ -7,5 +7,5 @@ import org.springframework.web.bind.annotation.PathVariable;
 @FeignClient("fraud")
 public interface FraudClient {
     @GetMapping(path = "api/v1/fraud-check/{customerId}")
-    FraudCheckResponse isFraudster(@PathVariable("customerId") Integer customerId);
+    FraudCheckResponse isFraudster(@PathVariable("customerId") Long customerId);
 }
